[move] SeichiAssistItemMigrationsのパッケージを移動する
